Remove Dup AES-GCM Code and Split Off OpenSSL code
This splits file skenc.cpp into OpenSSL-dependent and -independent parts.
It also removes multiple implementations of EncryptMessage().

- skenc.cpp
  - keep low-level OpenSSL-dependent code in this file
  - EncryptMessage(): remove duplicate implementations

- skenc_common.cpp (new file)
  - Move OpenSSL-independent code from skenc.cpp to here
  - GenerateIV(): use Avalon hash function instead of OpenSSL hash function

- Makefile
  - Clean up duplicate rules

This will make implementing skenc.cpp with MbedTLS easier.
